% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTest1.R
\name{BayesTest1}
\alias{BayesTest1}
\title{Bayesian One-Sample Tests for Covariance Matrix}
\usage{
BayesTest1(data, Sigma0 = diag(ncol(data)), method = c("mxPBF"), ...)
}
\description{
Bayesian One-Sample Tests for Covariance Matrix
}
