% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDiagTest1.mxPBF.R
\name{BDiagTest1.mxPBF}
\alias{BDiagTest1.mxPBF}
\title{One-Sample Diagonality Test by Maximum Pairwise Bayes Factor}
\usage{
BDiagTest1.mxPBF(data, a0 = 2, b0 = 2, gamma = 1)
}
\arguments{
\item{data}{an \eqn{(n\times p)} data matrix where each row is an observation.}

\item{a0}{shape parameter for inverse-gamma prior.}

\item{b0}{scale parameter for inverse-gamma prior.}

\item{gamma}{non-negative number. See the equation above.}
}
\value{
a named list containing: \describe{
\item{log.BF.mat}{ \eqn{(p\times p)} matrix of pairwise log Bayes factors.}
}
}
\description{
One-sample diagonality test can be stated with the null hypothesis
\deqn{H_0 : \sigma_{ij} = 0~\mathrm{for any}~i \neq j}
and alternative hypothesis \eqn{H_1 : ~\mathrm{not}~H_0}
with \eqn{\Sigma_n = (\sigma_{ij})}. Let \eqn{X_i} be the \eqn{i}-th column of data matrix. Under the maximum pairwise bayes factor framework, we have following hypothesis,
\deqn{H_0: a_{ij}=0\quad \mathrm{versus. } \quad  H_1: \mathrm{ not }~ H_0.}
The model is
\deqn{X_i | X_j \sim N_n( a_{ij}X_j, \tau_{ij}^2 I_n ).}
Under \eqn{H_0}, the prior is set as
\deqn{\tau_{ij}^2 \sim IG(a0, b0)} and under \eqn{H_1}, priors are
\deqn{ a_{ij}|\tau_{ij}^2 \sim N(0, \tau_{ij}^2/(\gamma*||X_j||^2))}
\deqn{\tau_{ij}^2 \sim IG(a0, b0).}
}
\examples{
\dontrun{
## generate data from multivariate normal with trivial covariance.
pdim = 10
data = matrix(rnorm(100*pdim), nrow=100)

## run test
## run mxPBF-based test
out1 = BDiagTest1.mxPBF(data)
out2 = BDiagTest1.mxPBF(data, a0=5.0, b0=5.0) # change some params

## visualize two Bayes Factor matrices
opar <- par(mfrow=c(1,2), pty="s")
image(exp(out1$log.BF.mat)[,pdim:1], main="default")
image(exp(out2$log.BF.mat)[,pdim:1], main="a0=b0=5.0")
par(opar)
}

}
\references{
\insertRef{lee_maximum_2018}{CovTools}
}
