% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-CovTools.R
\docType{package}
\name{CovTools-package}
\alias{CovTools-package}
\title{CovTools : A Collection of Geometric and Statistical Tools for Covariance Analysis}
\description{
Covariance is of universal prevalence across various disciplines within statistics.
The \pkg{CovTools} aims at providing a rich collection of geometric and inferential tools
for convenient analysis of covariance structures, topics including distance measures,
mean covariance estimator, covariance hypothesis test for one-sample and two-sample cases,
and covariance estimation.
}
