% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-CovTools.R
\docType{package}
\name{package-CovTools}
\alias{package-CovTools}
\title{A Collection of Geometric and Statistical Tools for Covariance (and Precision) Analysis}
\description{
Covariance is of universal prevalence across various disciplines within statistics.
\pkg{CovTools} package aims at providing a rich collection of geometric and statistical tools
for a variety of inferences on covariance structures as well as its inverse called precision matrix.
See the sections below for a comprehensive list of functions provided from the package.
}
\section{Geometric Methods}{

From inference on manifolds perspective, we have following functions,
\tabular{cc}{
\emph{name of a function} \tab \emph{description} \cr
\code{\link{CovDist}} \tab compute pairwise distance of covariance matrices \cr
\code{\link{CovMean}} \tab compute mean covariance matrix
}
}

